Add files via upload
Perl script for creating UCSC track hub with a list of bigWig / bigBed files. Creates a separate hub every time it is run. For creating multiple super tracks in a single hub, rerun the script with the various lists of files. The hub files are recreated, but they work as long as the genome and version are the same.